requirements.txt
numpy==1.18.1
Pillow==8.0.1
PyQt5==5.15.2
